+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/model/LightWalletEndpoint.kt
@@ -1,0 +1,5 @@
+package cash.z.ecc.android.sdk.model
+
+data class LightWalletEndpoint(val host: String, val port: Int, val isSecure: Boolean) {
+    companion object
+}