+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/model/LightwalletdServer.kt
@@ -1,5 +1,0 @@
-package cash.z.ecc.android.sdk.model
-
-data class LightwalletdServer(val host: String, val port: Int, val isSecure: Boolean) {
-    companion object
-}