+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Change Log
 
 Upcoming
 ------------------------------------
- - Split `ZcashNetwork` into `ZcashNetwork` and `LightwalletdServer` to decouple network and server configuration
+ - Split `ZcashNetwork` into `ZcashNetwork` and `LightWalletEndpoint` to decouple network and server configuration
 
 Version 1.8.0-beta01
 ------------------------------------